+++ src/main/java/com/spring/hotelspace/admin/management/user/dao/AdminManageClientDAO.java
@@ -1,5 +1,11 @@
 package com.spring.hotelspace.admin.management.user.dao;
 
+import java.util.List;
+
+import com.spring.hotelspace.admin.main.vo.AdminMainVO;
+import com.spring.hotelspace.admin.management.user.vo.AdminManageClientVO;
+
 public interface AdminManageClientDAO {
 
+	public List<AdminManageClientVO> getClientTable(AdminManageClientVO vo);
 }
