+++ src/main/java/com/spring/hotelspace/admin/management/user/service/AdminManageClientService.java
@@ -1,5 +1,10 @@
 package com.spring.hotelspace.admin.management.user.service;
 
+import java.util.List;
+
+import com.spring.hotelspace.admin.management.user.vo.AdminManageClientVO;
+
 public interface AdminManageClientService {
 
+	public List<AdminManageClientVO> getClientList(AdminManageClientVO vo);
 }
